Breakout lineage test into their own test class. Add UI test for lineage.
